backend/requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.8.2
click==8.1.8
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
PyJWT==2.9.0
PyMySQL==1.1.1
SQLAlchemy==2.0.38
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
alembic==1.14.1
bcrypt==4.2.1
blinker==1.8.2
click==8.1.8
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pika==1.3.2
PyJWT==2.9.0
PyMySQL==1.1.1
SQLAlchemy==2.0.38
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
alembic==1.14.1
bcrypt==4.2.1
blinker==1.8.2
click==8.1.8
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pika==1.3.2
PyJWT==2.9.0
PyMySQL==1.1.1
SQLAlchemy==2.0.38
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
